Restructure + refactor backend types
