+++ .gitignore
@@ -1,6 +1,8 @@
 build
 .DS_Store
 
+.vscode
+
 # Logs
 logs
 *.log
@@ -114,3 +116,6 @@ dist
 docs/latex/*.pdf
 docs/latex/*.gz
 docs/latex/images
+
+tsconfig.electron.prod.json
+tsconfig.svelte.prod.json
