--- src/electron/lib/api/IpcResponse.ts
@@ -1,4 +1,0 @@
-export interface IpcResponse<T> {
-  success: boolean;
-  data: T;
-}
