+++ src/electron/lib/api/apis/UtilApi.ts
@@ -2,8 +2,9 @@ import type { ElectronMainApi } from "electron-affinity/main";
 // import type { Blix } from "../Blix"
 
 import { platform, type, release } from "os";
-import logger from "../../utils/logger";
+import logger from "../../../utils/logger";
 
+// Exposes basic system information
 export class UtilApi implements ElectronMainApi<UtilApi> {
   private _counter = 0;
   // private _blix: Blix
