+++ src/electron/lib/core-graph/CoreGraphSubscriber.ts
