--- src/electron/lib/interfaces.ts
@@ -1,12 +1,0 @@
-// interface ILinear {
-//     multiplier: number | undefined;
-//     offset: number | undefined;
-// }
-
-export interface IEditPhoto {
-  brightness: number | undefined;
-  saturation: number | undefined;
-  hue: number | undefined;
-  rotate: number | undefined;
-  shadow: number | undefined;
-}
