+++ src/electron/lib/projects/ProjectManager.ts
@@ -1,14 +1,15 @@
 import { CoreProject } from "./CoreProject";
 import logger from "../../utils/logger";
-import type { IpcResponse } from "../api/IpcResponse";
 import { join } from "path";
 import { app } from "electron";
 import fs from "fs";
 import type { PathLike } from "fs";
 import type { UUID } from "../../../shared/utils/UniqueEntity";
-import type { MainWindow } from "../api/WindowApi";
+import type { MainWindow } from "../api/apis/WindowApi";
 import type { CommonProject } from "../../../shared/types";
 import { dialog } from "electron";
+import type { IpcResponse } from "../api/MainApi";
+
 // This should kinda be extending Registry and then called ProjectRegistry
 // instead of Project Manager but I don't feel like the Registry interface is
 // the correct fit for this.
