+++ src/electron/lib/registries/CommandRegistry.ts
@@ -1,4 +1,4 @@
-import type { Registry, RegistryInstance } from "../Registry";
+import type { Registry, RegistryInstance } from "./Registry";
 import type { ICommand } from "../../../shared/types";
 
 export class CommandRegistry implements Registry {
