+++ src/electron/lib/registries/Registry.ts
